Move the frontend logger to a separate file
Currently, when importing `hodlog`, it imports `bunyan`, which we don't
need when we're in the browser.

This change moves the frontend logger to its own file, so we can import
only that file when we're running the client, and bunyan won't end up in
the browser bundle.
