+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hodlog",
-  "version": "0.4.0",
+  "version": "0.4.1",
   "author": "Zoover",
   "main": "./build/index.js",
   "scripts": {
