+++ src/logger.js
@@ -2,28 +2,9 @@
 
 import bunyan from 'bunyan';
 import stdIOStream from './stdIOStream';
+import FrontEndLogger from './frontend';
 
 const logLevel = process.env.LOG_LEVEL || 'info';
-
-function FrontEndLogger() {
-  this.isFrontEnd = true;
-  this.info = console.log.bind(console); // eslint-disable-line no-console
-  this.error = function proxyError(...args) {
-    if (typeof Rollbar !== 'undefined' && typeof Rollbar.error === 'function') {
-      Rollbar.error(...args);
-    }
-
-    console.error(...args); // eslint-disable-line no-console
-  };
-  this.warn = function proxyWarn(...args) {
-    if (typeof Rollbar !== 'undefined' && typeof Rollbar.warning === 'function') {
-      Rollbar.warning(...args);
-    }
-
-    console.warn(...args); // eslint-disable-line no-console
-  };
-  this.trace = console.trace.bind(console); // eslint-disable-line no-console
-}
 
 export default (function logger() {
   if (typeof window !== 'undefined') {
